Struct mouse_data::MouseButtons

pub struct MouseButtons { /* private fields */ }

Implementations

impl MouseButtons

pub const fn new() -> Self

Returns an instance with zero initialized data.

impl MouseButtons

pub const fn into_bytes(self) -> [u8; 1]

Returns the underlying bits.

Layout

The returned byte array is layed out in the same way as described here.

pub const fn from_bytes(bytes: [u8; 1]) -> Self

Converts the given bytes directly into the bitfield struct.

impl MouseButtons

pub fn left(&self) -> <bool as Specifier>::InOut

Returns the value of left.

pub fn left_or_err(
&self
) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of left.

#Errors

If the returned value contains an invalid bit pattern for left.

pub fn with_left(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of left set to the given value.

#Panics

If the given value is out of bounds for left.

pub fn with_left_checked(
self,
new_val: <bool as Specifier>::InOut
) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of left set to the given value.

#Errors

If the given value is out of bounds for left.

pub fn set_left(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of left to the given value.

#Panics

If the given value is out of bounds for left.

pub fn set_left_checked(
&mut self,
new_val: <bool as Specifier>::InOut
) -> Result<(), OutOfBounds>

Sets the value of left to the given value.

#Errors

If the given value is out of bounds for left.

pub fn right(&self) -> <bool as Specifier>::InOut

Returns the value of right.

pub fn right_or_err(
&self
) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of right.

#Errors

If the returned value contains an invalid bit pattern for right.

pub fn with_right(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of right set to the given value.

#Panics

If the given value is out of bounds for right.

pub fn with_right_checked(
self,
new_val: <bool as Specifier>::InOut
) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of right set to the given value.

#Errors

If the given value is out of bounds for right.

pub fn set_right(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of right to the given value.

#Panics

If the given value is out of bounds for right.

pub fn set_right_checked(
&mut self,
new_val: <bool as Specifier>::InOut
) -> Result<(), OutOfBounds>

Sets the value of right to the given value.

#Errors

If the given value is out of bounds for right.

pub fn middle(&self) -> <bool as Specifier>::InOut

Returns the value of middle.

pub fn middle_or_err(
&self
) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of middle.

#Errors

If the returned value contains an invalid bit pattern for middle.

pub fn with_middle(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of middle set to the given value.

#Panics

If the given value is out of bounds for middle.

pub fn with_middle_checked(
self,
new_val: <bool as Specifier>::InOut
) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of middle set to the given value.

#Errors

If the given value is out of bounds for middle.

pub fn set_middle(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of middle to the given value.

#Panics

If the given value is out of bounds for middle.

pub fn set_middle_checked(
&mut self,
new_val: <bool as Specifier>::InOut
) -> Result<(), OutOfBounds>

Sets the value of middle to the given value.

#Errors

If the given value is out of bounds for middle.

pub fn fourth(&self) -> <bool as Specifier>::InOut

Returns the value of fourth.

pub fn fourth_or_err(
&self
) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of fourth.

#Errors

If the returned value contains an invalid bit pattern for fourth.

pub fn with_fourth(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of fourth set to the given value.

#Panics

If the given value is out of bounds for fourth.

pub fn with_fourth_checked(
self,
new_val: <bool as Specifier>::InOut
) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of fourth set to the given value.

#Errors

If the given value is out of bounds for fourth.

pub fn set_fourth(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of fourth to the given value.

#Panics

If the given value is out of bounds for fourth.

pub fn set_fourth_checked(
&mut self,
new_val: <bool as Specifier>::InOut
) -> Result<(), OutOfBounds>

Sets the value of fourth to the given value.

#Errors

If the given value is out of bounds for fourth.

pub fn fifth(&self) -> <bool as Specifier>::InOut

Returns the value of fifth.

pub fn fifth_or_err(
&self
) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of fifth.

#Errors

If the returned value contains an invalid bit pattern for fifth.

pub fn with_fifth(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of fifth set to the given value.

#Panics

If the given value is out of bounds for fifth.

pub fn with_fifth_checked(
self,
new_val: <bool as Specifier>::InOut
) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of fifth set to the given value.

#Errors

If the given value is out of bounds for fifth.

pub fn set_fifth(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of fifth to the given value.

#Panics

If the given value is out of bounds for fifth.

pub fn set_fifth_checked(
&mut self,
new_val: <bool as Specifier>::InOut
) -> Result<(), OutOfBounds>

Sets the value of fifth to the given value.

#Errors

If the given value is out of bounds for fifth.